requirements.txt
appdirs==1.4.4
asgiref==3.3.4
cfgv==3.3.0
distlib==0.3.2
Django==3.2.3
djangorestframework==3.12.4
filelock==3.0.12
identify==2.2.6
importlib-metadata==4.3.1
nodeenv==1.6.0
pkg-resources==0.0.0
pre-commit==2.13.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.4.7
zipp==3.4.1
